Refactor server.py into more specific files under pages/. (#1227)
* progress

* Refactor server.py into more specific files under pages/.

* remove server_test.py since tests are in specific files.

* Fix flakey test (feature ID could match part of date)
